Add set_map_localtion to TensorPipeOptions to support GPU args
This commit enables sending non-CPU tensors through RPC using
TensorPipe backend. Users can configure device mappings by calling
`set_map_location` on `TensorPipeRpcBackendOptions`. Internally,
the `rpc_init` API verifies the correctness of device mappings. It
will shutdown RPC if the check failed, or proceed and pass global
mappings to TensorPipeAgent if the check was successful. For serde,
we added a device indices field to TensorPipe read and write buffers,
which should be either empty (all tensors must be on CPU) or match
the tensors in order and number in the RPC message. This commit
does not yet avoid zero-copy, the tensor is always moved to CPU
on the sender and then moved to the specified device on the receiver.

ghstack-source-id: 4ced68d26496a22aebb81a558a48da34000ab4d7
Pull Request resolved: https://github.com/pytorch/pytorch/pull/42637